+++ WebApp/Controllers/KeyIndicatorsController.cs
@@ -191,7 +191,7 @@ namespace WebApp.Controllers
             try
             {
                 objOnlyOnceTxn = KR.GetOnlyOnceTxnData(userDetails.GroupId, outletId, type, userDetails.connectionString, userDetails.LoginId);
-                long? totalspent = 0;
+                decimal? totalspent = 0;
                 long? totalvisit = 0;
                 long? availbal = 0;
                 foreach (var item in objOnlyOnceTxn)
@@ -372,7 +372,7 @@ namespace WebApp.Controllers
             try
             {
                 objNonTransactingTxn = KR.GetNonTransactingTxnData(userDetails.GroupId, outletId, barType, userDetails.connectionString, userDetails.LoginId);
-                long? totalspent = 0;
+                decimal? totalspent = 0;
                 long? totalvisit = 0;
                 long? availbal = 0;
                 foreach (var item in objNonTransactingTxn)
